Major overhaul. Using regex, NGG codon avoidance, big bugfix.
There is a LOT of cruft in this update, but a lot of good too. For
starters, it has a huge bugfix: I'd forgotten to include reverse
complement in the search process for sites-to-exclude. Improving
on search even more, I've included regex-conversion from IUPAC
notation, and the search system now uses regex to find troubling
sites to address, rather than creating huge lists of possible combinations
of IUPAC-denoted characters.

This means this revision of pysplicer should be able to handle exclusions
like ACGNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNGG, whereas prior versions
would choke trying to expand that into a set of comparable strings.

This revision also includes widespread method-isation of subfunctions,
to allow better cross-use of code. For example, normalisation of codon
tables used to be handled by a function within the codon table import
functions, but on-the-fly codon choice adaptation is better served with
a method that does this, so that's been implemented.

Finally, building on the preceding, there is now a default rule that
attempts to avoid NGG codons within the first 25 codons, based on research
that suggests this can be a cause of premature ribosomal dropoff.

There's a lot here that needs to be deleted, and hadn't yet been.